feat(android): expose windowFocusChanged as a command
Some versions of Unity Player show black screen after remounted (see #59
for reference). The root cause is not clear but putting the app in
background and then bringing it back fixes the problem. To simulate this
process, `UnityPlayer.windowFocusChanged` is exposed as a command to be
invoked from within React Native when necessary.